+++ components/core/typo/Text.mjs
@@ -16,9 +16,10 @@ export default class Text extends AbstractComponent {
      * @param {string} theme Component color theme
      * @param {string} id Component ID
      * @param {string[]} classes Component classes
+     * @param {object} attributes Component attributes
      */
-    constructor(props = {}, children = [], theme = null, id = null, classes = []) {
-        super(props, children, theme, id, classes);
+    constructor(props = {}, children = [], theme = null, id = null, classes = [], attributes = {}) {
+        super(props, children, theme, id, classes, attributes);
     }
 
     static sizes = {
@@ -30,6 +31,14 @@ export default class Text extends AbstractComponent {
         md: 'md',
         s: 's'
     }
+
+    static textThemeClass(theme) { if (!Object.values(this._themes).includes(theme)) { console.warn(`${this.constructor.name} (${this.id}): Text theme variant '${theme}' is not recognized`) }; return `text-${theme}`; }
+    static sizeClass(size) { if (!Object.values(this.sizes).includes(size)) { console.warn(`${this.constructor.name} (${this.id}): Text size variant '${size}' is not recognized`) }; return `text-${size}`; }
+    static fontClass(font) { if (!Object.values(this.fonts).includes(font)) { console.warn(`${this.constructor.name} (${this.id}): Text font variant '${font}' is not recognized`) }; return `text-${font}`; }
+    static alignmentClass(alignment) { if (!Object.values(this.alignments).includes(alignment)) { console.warn(`${this.constructor.name} (${this.id}): Text alignment variant '${alignment}' is not recognized`) }; return `text-${alignment}`; }
+    static mutedClass() { return `text-muted`; }
+    static italicClass() { return `text-italic`; }
+    static strongClass() { return `text-strong`; }
 
     static fonts = {
         sans: 'sans',
@@ -69,12 +78,13 @@ export default class Text extends AbstractComponent {
     render() {
         let text = document.createElement(this._elementTagName ?? "span");
 
-        if (this.props.size) text.classList.add(`text-${this.props.size}`);
-        if (this.props.font) text.classList.add(`text-${this.props.font}`);
-        if (this.props.align) text.classList.add(`text-${this.props.align}`);
-        if (this.props.muted) text.classList.add(`text-muted`);
-        if (this.props.italic) text.classList.add(`text-italic`);
-        if (this.props.bold) text.classList.add(`text-strong`);
+        if (Object.values(this.constructor._themes).includes(this.theme)) text.classList.add(this.constructor.textThemeClass(this.theme));
+        if (this.props.size) text.classList.add(this.constructor.sizeClass(this.props.size));
+        if (this.props.font) text.classList.add(this.constructor.fontClass(this.props.font));
+        if (this.props.align) text.classList.add(this.constructor.alignmentClass(this.props.align));
+        if (this.props.muted) text.classList.add(this.constructor.mutedClass());
+        if (this.props.italic) text.classList.add(this.constructor.italicClass());
+        if (this.props.bold) text.classList.add(this.constructor.strongClass());
 
         this.attachChildren(text);
         this.attachListeners(text);
@@ -86,12 +96,49 @@ export default class Text extends AbstractComponent {
         if (this.isMounted() && JSON.stringify(this._props) !== JSON.stringify(oldProps)) {
             let el = this.i();
 
-            if (oldProps.size !== this.props.size) el.classList.remove(`text-${oldProps.size}`); if (this.props.size) el.classList.add(`text-${this.props.size}`);
-            if (oldProps.font !== this.props.font) el.classList.remove(`text-${oldProps.font}`); if (this.props.font) el.classList.add(`text-${this.props.font}`);
-            if (oldProps.align !== this.props.align) el.classList.remove(`text-${oldProps.align}`); if (this.props.align) el.classList.add(`text-${this.props.align}`);
-            oldProps.muted !== this.props.muted && this.props.muted ? el.classList.add(`text-muted`) : el.classList.remove(`text-muted`);
-            oldProps.italic !== this.props.italic && this.props.italic ? el.classList.add(`text-italic`) : el.classList.remove(`text-italic`);
-            oldProps.bold !== this.props.bold && this.props.bold ? el.classList.add(`text-strong`) : el.classList.remove(`text-strong`);
+
+            if (oldProps.size !== this.props.size) {
+                el.classList.remove(this.constructor.sizeClass(oldProps.size));
+
+                if (this.props.size) el.classList.add(this.constructor.sizeClass(this.props.size));
+            }
+            if (oldProps.font !== this.props.font) {
+                el.classList.remove(this.constructor.fontClass(oldProps.font));
+
+                if (this.props.font) el.classList.add(this.constructor.fontClass(this.props.font));
+            }
+            if (oldProps.align !== this.props.align) {
+                el.classList.remove(this.constructor.alignmentClass(oldProps.align));
+
+                if (this.props.align) el.classList.add(this.constructor.alignmentClass(this.props.align));
+            }
+            if (oldProps.muted !== this.props.muted) {
+                el.classList.remove(this.constructor.mutedClass());
+
+                if (this.props.muted) el.classList.add(this.constructor.mutedClass());
+            }
+            if (oldProps.italic !== this.props.italic) {
+                el.classList.remove(this.constructor.italicClass());
+
+                if (this.props.italic) el.classList.add(this.constructor.italicClass());
+            }
+            if (oldProps.bold !== this.props.bold) {
+                el.classList.remove(this.constructor.strongClass());
+
+                if (this.props.bold) el.classList.add(this.constructor.strongClass());
+            }
+        }
+    }
+
+    onThemeChange(oldTheme) {
+        super.onThemeChange(oldTheme);
+
+        if (this.isMounted() && oldTheme !== this.theme) {
+            let i = this.i();
+
+            i.classList.remove(this.constructor.textThemeClass(oldTheme));
+
+            if (Object.values(this.constructor._themes).includes(this.theme)) i.classList.add(this.constructor.textThemeClass(this.theme));
         }
     }
 }