+++ components/display/Tooltip.mjs
@@ -7,14 +7,17 @@ export default class Tooltip extends AbstractComponent {
     /**
      * Creates a new component
      * @param {object} props Component properties
+     * @param {string|null} props.content
      * @param {string} props.location
+     * @param {boolean} props.inline
      * @param {string[]|HTMLElement[]|AbstractElement[]} children Component children
      * @param {string} theme Component color theme
      * @param {string} id Component ID
      * @param {string[]} classes Component classes
+     * @param {object} attributes Component attributes
      */
-    constructor(props = {}, children = [], theme = null, id = null, classes = []) {
-        super(props, children, theme, id, classes);
+    constructor(props = {}, children = [], theme = null, id = null, classes = [], attributes = {}) {
+        super(props, children, theme, id, classes, attributes);
     }
 
     static locations = {