+++ components/layout/accordion/AccordionPlusMinusSummary.mjs
@@ -14,9 +14,11 @@ export default class AccordionPlusMinusSummary extends AccordionSummary {
      * @param {string[]|HTMLElement[]|AbstractElement[]} children Component children
      * @param {string} theme Component color theme
      * @param {string} id Component ID
+     * @param {string[]} classes Component classes
+     * @param {object} attributes Component attributes
      */
-    constructor(props = {}, children = [], theme = null, id = null) {
-        super(props, children, theme, id);
+    constructor(props = {}, children = [], theme = null, id = null, classes = [], attributes = {}) {
+        super(props, children, theme, id, classes, attributes);
 
         this._icon = new FontawesomeIcon({
             name: "plus"