+++ components/layout/dialog/Modal.mjs
@@ -15,9 +15,11 @@ export default class Modal extends Dialog {
      * @param {string[]|HTMLElement[]|AbstractElement[]} children Component children
      * @param {string} theme Component color theme
      * @param {string} id Component ID
+     * @param {string[]} classes Component classes
+     * @param {object} attributes Component attributes
      */
-    constructor(props = {}, children = [], theme = null, id = null) {
-        super(props, children, theme, id);
+    constructor(props = {}, children = [], theme = null, id = null, classes = [], attributes = {}) {
+        super(props, children, theme, id, classes, attributes);
 
         this._closeBtn = new Button({
             style: Button.styles.empty,